Day 6/60 - The Decentralized Semantic Web Mirror
Thinking ahead, what will be needed to be built infrastructure-wise for a system like the one I want to make? The answer: Web 3.... (the semantic web).. So now to try to solve a decade old problem of re-indexing the web. As startup costs are a thing, hoping to just serve as a CDN of sorts and outsource the processing and storage to the community. Ideating on this concept.